// src/api/openapi/model/companyEvent.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserEssential } from './userEssential';
import { CompanyEventEnum } from './companyEventEnum';


export interface CompanyEvent { 
    title: any | null;
    date: any | null;
    event_type: CompanyEventEnum;
    id: any | null;
    timestamp: any | null;
    user: UserEssential;
}
export namespace CompanyEvent {
}